fix: fixing coding snippets for reference storybook docs [TOL-1398]
