+++ packages/reference/stories/MultipleEntryReferenceEditor/IMPORT_SNIPPET.md
@@ -1,0 +1,3 @@
+```jsx
+import { MultipleEntryReferenceEditor } from '@contentful/field-editor-reference';
+```
