+++ packages/reference/stories/MultipleMediaEditor/IMPORT_SNIPPET.md
@@ -1,0 +1,3 @@
+```jsx
+import { MultipleMediaEditor } from '@contentful/field-editor-reference';
+```
