+++ packages/reference/stories/SingleEntryReferenceEditor/FIELD_APP_SNIPPET.md
@@ -1,0 +1,13 @@
+```js
+import { SingleEntryReferenceEditor } from '@contentful/field-editor-reference';
+
+/// your app code
+init((sdk) => {
+  if (sdk.location.is(locations.LOCATION_ENTRY_FIELD)) {
+    render(
+      <SingleEntryReferenceEditor viewType="card" sdk={sdk} />,
+      document.getElementById('root')
+    );
+  }
+});
+```
