+++ packages/reference/stories/SingleEntryReferenceEditor/IMPORT_SNIPPET.md
@@ -1,0 +1,3 @@
+```jsx
+import { SingleEntryReferenceEditor } from '@contentful/field-editor-reference';
+```
