+++ packages/reference/stories/SingleMediaEditor/IMPORT_SNIPPET.md
@@ -1,0 +1,3 @@
+```jsx
+import { SingleMediaEditor } from '@contentful/field-editor-reference';
+```
